refactor: move UIntPtr32<T> to new file
